@Monzey review part 2 - Move state management to parent
