+++ src/Molecules/ActionCard/ActionCard.jsx
@@ -1,14 +1,7 @@
 import React from 'react'
 
-import { Times } from '@styled-icons/fa-solid/Times'
-import { Check } from 'styled-icons/material/Check'
-import { Edit } from 'styled-icons/material/Edit'
-
 import Card from '../../Atoms/Card'
-import Trans from '../../Atoms/Trans'
 import FlexBox from '../../Templates/FlexBox'
-
-let originalValue
 
 const Action = props => <FlexBox
   css={{ cursor: 'pointer' }}
@@ -17,6 +10,7 @@ const Action = props => <FlexBox
   alignItems="center"
   border="none"
   mt={4}
+  justifyContent="space-between"
   {...props}
 />
 
@@ -33,53 +27,11 @@ const Description = props => <FlexBox
   {...props}
 />
 
-const Actions = ({
-  edit,
-  setEdit,
-  value,
-  setValue,
-}) => {
-  const onCancel = () => {
-    setValue(originalValue)
-    setEdit(false)
-  }
-
-  const onEdit = () => {
-    originalValue = value
-    setEdit(true)
-  }
-
-  const onSave = () => {
-    setEdit(false)
-  }
-
-  return edit ? (
-    <>
-      <Action onClick={onCancel}>
-        <Times size={16} />
-        <FlexBox marginLeft={0.5}><Trans>global.action.cancel</Trans></FlexBox>
-      </Action>
-      <Action onClick={onSave}>
-        <Check size={16} />
-        <FlexBox marginLeft={0.5}><Trans>global.action.save</Trans></FlexBox>
-      </Action>
-    </>
-  ) : (
-    <Action onClick={onEdit}>
-      <Edit size={16} />
-      <FlexBox marginLeft={0.5}><Trans>global.action.edit</Trans></FlexBox>
-    </Action>
-  )
-}
-
 const ActionCard = ({
   title,
   description,
+  actions,
   children,
-  setEdit,
-  edit = false,
-  value,
-  setValue,
   ...props
 }) => {
   return (
@@ -91,9 +43,9 @@ const ActionCard = ({
         {description}
       </Description>
       {children}
-      <FlexBox justifyContent="space-between">
-        <Actions edit={edit} setEdit={setEdit} value={value} setValue={setValue} />
-      </FlexBox>
+      <Action>
+        {actions}
+      </Action>
     </Card>
   )
 }
