+++ src/Molecules/ActionCard/ActionCard.test.jsx
@@ -4,8 +4,14 @@ import '@testing-library/jest-dom/extend-expect'
 
 import ActionCard from './index'
 import Editable from '../Editable'
+import FlexBox from '../../Templates/FlexBox'
+
 import { text } from '@storybook/addon-knobs'
 import userEvent from '@testing-library/user-event'
+import { Times } from '@styled-icons/fa-solid/Times'
+import { Check } from 'styled-icons/material/Check'
+import { Edit } from 'styled-icons/material/Edit'
+import Trans from '../../Atoms/Trans'
 
 const title = 'This is a title'
 const description = 'This is a description'
@@ -16,18 +22,62 @@ const cancelBtn = 'global.action.cancel'
 const editBtn = 'global.action.edit'
 const saveBtn = 'global.action.save'
 
+let originalValue
+
+const Actions = ({
+  edit,
+  onEdit,
+  value,
+  onChange,
+}) => {
+  const onCancelClick = () => {
+    onChange(originalValue)
+    onEdit(false)
+  }
+
+  const onEditClick = () => {
+    originalValue = value
+    onEdit(true)
+  }
+
+  const onSaveClick = () => {
+    onEdit(false)
+  }
+
+  return edit ? (
+    <>
+      <FlexBox onClick={onCancelClick}>
+        <Times size={16} />
+        <FlexBox marginLeft={0.5}><Trans>global.action.cancel</Trans></FlexBox>
+      </FlexBox>
+      <FlexBox onClick={onSaveClick}>
+        <Check size={16} />
+        <FlexBox marginLeft={0.5}><Trans>global.action.save</Trans></FlexBox>
+      </FlexBox>
+    </>
+  ) : (
+    <FlexBox onClick={onEditClick}>
+      <Edit size={16} />
+      <FlexBox marginLeft={0.5}><Trans>global.action.edit</Trans></FlexBox>
+    </FlexBox>
+  )
+}
+
 const ActionCardWrapper = () => {
   const [edit, setEdit] = useState(false)
   const [value, setValue] = useState('Text #1')
+  const actions = <Actions
+    edit={edit}
+    onEdit={setEdit}
+    value={value}
+    onChange={setValue}
+  />
 
   return (
     <ActionCard
       description={description}
       title={title}
-      value={value}
-      setValue={setValue}
-      edit={edit}
-      setEdit={setEdit}
+      actions={actions}
     >
       <Editable
         edit={edit}
